add token authentication for export
